refactor: flytt ProsesseringInfoProvider og TokenValidationContextHolder fra ApplicationConfig
